+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The changelog for [KommunicateChatUI-iOS-SDK](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK). Also see the [releases](https://github.com/Kommunicate-io/KommunicateChatUI-iOS-SDK/releases) on Github.
 
+## [1.2.1] 2023-11-02
+- Fixed iOS 17 Button issue.
+- Fixed the Typing Customization issue.
+- Fixed buttons are getting cut in form template
+- Added hidepostCTA support for all types of buttons.
+- Default configuration added for hiding the form submit button with 'hidePostFormSubmit'.
+
 ## [1.2.0] 2023-10-27
 - Fixed Button Spacing